requirements.txt
pytest
pyyaml
pytest-django